Add libpq_be function to report bytes pending send in internal buffer
We have pq_is_send_pending() to report whether there's anything waiting
in the libpq send buffer, but not how much.

For stats reporting, add pq_nbytes_pending() so we can see how much is waiting
to send/receive. The send buffer is the main one of interest, as in general the
receive buffer is small and transient, but it can still matter for some
workloads like logical receivers.