Add INTERRUPTIBLE option to VACUUM & ANALYZE, test interruption.
Partially because that's practically useful, partially because it's
useful for testing (see 5871f09c985).

To implement add the new PROC_IS_INTERRUPTIBLE flag, controlling
whether a process should be interrupted if it holds a required
lock. As PROC_VACUUM_FOR_WRAPAROUND was only used to signal that a
PROC_IS_AUTOVACUUM proc should not be interrupted, I have removed
that.

Now that it's easier to trigger interruptible vacuums, add a test
doing so.

FIXME: change commit message based on whether we keep the behaviour of
autovacuum cancelling analyze or not.

Discussion: https://www.postgresql.org/message-id/20200909041147.amcitkxmlexgjfty%40alap3.anarazel.de