Add `truncate` option to subscription commands
When adding new tables to a subscription (either via `CREATE SUBSCRIPTION` or `REFRESH
PUBLICATION`), tables on the target which are being newly subscribed will be truncated before the
data copy step.  This saves explicit coordination of a manual `TRUNCATE` on the target tables and
allows the results of the initial data sync to be the same as on the publisher at the time of sync.

To preserve compatibility with existing behavior, the default value for this parameter is `false`.