Add table am 'tid_visible'
We directly call the heap function VM_ALL_VISIBLE in the
IndexOnlyNext function. This is not in line with the design idea of
table am. If the new storage type needs to implement index only
scan, he must hack the IndexOnlyNext function.

So this patch add a new table am 'tid_visible' to test visibility
of tid.

Co-authored-by: Jinbao Chen <cjinbao@vmware.com>
Co-authored-by: Asim R P <pasim@vmware.com>