Subqueries with LIMIT can be parallel safe when executing per-outer tuple
The code that determined whether or not a rel should be considered for
parallel query excluded subqueries with LIMIT/OFFSET. That's correct in
the general case: as the comment notes that'd mean we have to guarantee
ordering (and claims it's not worth checking that) for results to be
consistent across workers. However there's a simpler case to special
case than known unique or consistently ordered results: when we're
already going to execute the subquery within the context of each outer
tuple then whether we do that repeated execution within a single process
or multiple processes isn't going to affect the guarantees we offer
about consistency of results.