trace wal_level change for backup management
This patch adds a new parameter in the control file
in order to trace wal_level drop to minimal from higher level.
The purpose is to expose the LSN when the drop of wal_level
happens and to make it possible to notify the change to the users.
The value is reset when pg_basebackup creates a new base backup successfully.

Author: Takamichi Osumi <osumi.takamichi@fujitsu.com>
Discussion : https://www.postgresql.org/message-id/OSBPR01MB4888B94DF1C9325033C26D00EDD00%40OSBPR01MB4888.jpnprd01.prod.outlook.com