DROP TABLE breaks sync worker relid.
Previously we were allowing to drop the table when logical replication table synchronization was still in progress. This caused the sync worker's relid to become invalid, and then any subsequent use of that relid would cause problems. e.g. In the test case which exposed this bug, the problem manifested as a stacktrace when the sync worker LOG message attempted to display the table name using the dropped relid.