skip cci
In RI check, we currently call CommandCounterIncrement every time we insert into
a table with foreign key, which is not supported in a parallel worker. However, it's necessary
to do CCI only if the referenced table is modified during an INSERT command.

For now, all the cases that will modify the referenced table are treated as parallel unsafe.

We can skip CCI to let the RI check (for now only RI_FKey_check_ins) to be executed in a parallel worker.

For the parallel safety of RI_FKey_check_ins,
it internally uses [select pk rel for key share] which is considered as parallel unsafe in planner.
However, [select for key share] does not generate combo-cid or new command id.
And it acquires share lock which does not conflict with itself and works as expected with group locking.
In addition, as we assumed that the pk rel will not be modified, dead lock on pk rel will not happen.

Based on above, IMO, execute [select for key share] in parallel worker is as safe as parallel insert.