extend safery check to support parallel insert into fk relation
Currently, we cannot support parallel insert into a fk relation in all cases.

For example:
When inserting into a table with a foreign key, if the referenced table can also be modified by
the INSERT command, we will need to do CommandCounterIncrement to let the modification
on the referenced table be visible for the RI check, which is not supported in a parallel worker.

So, extend the parallel-safety check to treat the following cases (could modify referenced table) as parallel restricted:
1) foreign key and primary key are on the same table(INSERT's target table).
  (referenced and referencing are the same table)
2) referenced and referencing table are both partition of INSERT's target table.

(Note: modifyingCTE and function with modifying statement could also modifiy the referenced table,
However, the current parallel safety checks already treat it as unsafe, so we do not need to
do anything about it.)

Except for the above cases, we treat other cases as parallel safe.