Currently the set_append_rel_size doesn't consider the init partition
prune, so the estimated size may be wrong at a big scale sometimes.
In this patch I used the set the rows = parentrel->tuples *
clauseselecitivty. In this case we can loss some lose some accuracy when initial
partition prune doesn't happen at all. but generally I think it would be OK.

Another strategy is we should check if init partiton prune can happen.
if we are sure about that, we adapt the above way. or else we can use
the local stats strategy still.