disallow TRUNCATE on user_catalog_table
When TRUNCATE command is executed on user_catalog_table
in synchronous_mode during logial replication, TRUNCATE waits for
the subscriber's synchronization. At the same time, we don't have
a restriction that output plugin cannot take a lock on it. Therefore,
the walsender can hang due to the deadlock to take a new lock
on a user_catalog_table, since the table is locked by TRUNCATE already.
This patch addresses such a deadlock hazards by prohibiting
an operation to issue TRUNCATE on user_catalog_table.