Use (void) when return value of fsm_set_and_search is ignored
It looks like for some of the fsm_set_and_search calls whose
return value is ignored (in fsm_search and RecordPageWithFreeSpace),
there's no (void) casting. In the code base, we generally use (void)
when non-void return value of a function is ignored.

Although it is not a bug or not giving any compiler warning, just
use (void) as it seems to be a standard practice in the code base.