add the not null attrs for RelOptInfo. Here is how it works.
For baserel, it records the notnull attrs as a bitmapset and store it to
RelOptInfo->notnull_attrs[0].  As for the joinrel, suppose the relids is {1,3,
5}, then the notnull_attrs[1/3/5] will be used to store notnull_attrs bitmapset
for relation 1,3,5 separately. I don't handle this stuff for all kinds of upper
relation and subquery so far since it doesn't pass the design review yet.