Use a WaitLatch for Pre and Post Auth Delay
Instead of using pg_usleep() for waiting PostAuthDelay and
PreAuthDelay, use a WaitLatch. This has the advantage that
we will realize if the postmaster has been killed since the
last time we decided to sleep.

For PostAuthDelay, don't use WL_LATCH_SET. Because MyLatch, on
which WaitLatch going to wait, could have been set in
SwitchToSharedLatch() by InitProcess(). For PreAuthDelay, although
there's no problem to use WL_LATCH_SET as MyLatch still points to local
latch in BackendInitialize(), let's not use it as the MyLatch may
point to shared latch later.