No longer ignoring failures on file close
When writing the changes from transactions as part of logical
replication, the transient files to which transaction changes are
written were being closed without checking the return value from
close().  Fix that.

The return code from write() was already being handled reasonably,
but on filesystems which buffer write()s, the attempt to flush the
buffer during the close() operation may fail, with write() returning
-1 and setting errno.  This was not checked, meaning that some or
all of the changes for a transaction could be partially or entirely
missing, resulting in garbled data that cannot be replayed on the
subscriber side, or worse still, silently missing changes.