Use a WaitLatch for {pre, post}_auth_delay
Instead of using pg_usleep() for waiting PostAuthDelay and
PreAuthDelay, use a WaitLatch. This has the advantage that we
will realize if the postmaster has been killed since the last
time we decided to sleep.

For PostAuthDelay, don't use WL_LATCH_SET. Because MyLatch, on
which WaitLatch is going to wait, could have been set in
SwitchToSharedLatch() by InitProcess(). For PreAuthDelay, although
there is no problem to use WL_LATCH_SET as MyLatch still points to
local latch in BackendInitialize() and isn't fully initialized
for the backend. For this reason and  to be consistent with
PostAuthDelay let's not use WL_LATCH_SET for PreAuthDelay.