Partition key update in foreign table
Implements update of partition key for rows in foreign tables via foreign
delete + insert. This requires additional FDW apis: BeginForeignDelete
and EndForeignDelete, mainly because BeginForeignModify is, in any case,
called in the beginning of foreign update and we only know that we will
need separate remote delete+insert at ExecRun after checking partition
constraints. So postgresBeginForeignDelete keeps a separate state in
aux_delete_fmstate not to conflict with BeginForeignModify.