Cache parallel dml safety
The planner is updated to perform additional parallel-safety checks For
non-partitioned table if pg_class.relparalleldml is DEFAULT('d'), and cache the
parallel safety for the relation.

Whenever any function's parallel-safety is changed, invalidate the cached
parallel-safety for all relations in relcache for a particular database.

For partitioned table, If pg_class.relparalleldml is DEFAULT('d'), assume that
the table is UNSAFE to be modified in parallel mode.

If pg_class.relparalleldml is SAFE/RESTRICTED/UNSAFE, respect the specified
parallel dml safety instead of checking it again.