Improve Var op Var estimates with the same Var
When estimating (Var op Var) clauses, we can treat the case with the
same Var on both sides as a special case, and we can provide better
selectivity estimate than for the generic case.

For example for (a = a) clause we know it's (1.0 - nullfrac), because
all non-NULL rows are expected to match. Similarly for (a != a), wich
has selectivity 0.0. And the same reasoning can be applied to inequality
comparisons, like (a < a) etc.

This of course depends on the operators having the usual meaning (e.g.
"=" equality) and not some custom operator with different behavior, but
if they use the same estimator it's reasonable to assume that (we
already do that in various other places).

In principle, such clauses are a bit strange and queries are unlikely to
use them very often. But query generators sometimes do this, and these
checks are quite cheap so it's likely a win.