Fix pg_dump handling of ALTER DEFAULT PRIVILEGES IN SCHEMA.
When dumping ACLs, pg_dump uses the hard-wired default privileges
(i.e., acldefault()) as a reference.  While this is usually okay,
it sometimes doesn't work for schema-local default ACLs (i.e.,
pg_default_acl entries with defaclnamespace != 0).  This is because
the default for a schema-local default ACL entry is actually an
empty ACL, as per the following note in aclchk.c:

	/*
	 * The default for a global entry is the hard-wired default ACL for the
	 * particular object type.  The default for non-global entries is an empty
	 * ACL.  This must be so because global entries replace the hard-wired
	 * defaults, while others are added on.
	 */

To fix, adjust the pg_dump query for dumping ACLs to use NULL
instead of acldefault() for schema-local default ACLs.