Allow regular users to create event triggers
This is made safe by conditioning event triggers on the owning role
and making them fire only when the trigger owner is a direct or
indirect member of the role which performed the event.  As such, an
event trigger owner cannot trap users into performing actions unless
those users are ones the event owner could change role to anyway.

For superuser owners, nothing changes, because they can change role
to anybody and hence their event triggers still fire for everybody.
For non-superuser owners, the behavior does change, but
non-superuser owners of event triggers seem a sufficiently unusual
situation as to not need to preserve the old behavior.