Change log_(dis)connections to PGC_SIGHUP
log_connections is not effective when it is given in connection
options. Since no complaint has been heard for this behavior the
use-case looks rather thin.  Thus we change it to PGC_SIGHUP, rahther
than putting efforts to make it effective for the
use-case. log_disconnections is working with the usage but be
consistent by treating it the same way with log_connection.