unexport is_member_of_role(), add can_set_role()
is_member_of_role should not be used for privilege checking, so
to avoid that remove it from the acl header. The only remaining
usage is for SET ROLE in variable.c where membership is the
legitimate check, so a new exported explicetly named
can_set_role() is added in it's place.

Signed-off-by: Joshua Brindle <joshua.brindle@crunchydata.com>