document examples of setting up multiple values for shared_preload_libraries
The users can try different ways to set multiple values for
shared_preload_libraries even after reading the existing
documentation which is missing an example to show the right way.
The problem with the wrong parameter set command is that the ALTER
SYSTEM SET will not fail, but the server will not come up in case
it is restarted.

This patch adds few different ways one can set multiple values to
shared_preload_libraries.