pg_ident: allow delegation to an LDAP server
This is work-in-progress code, with several deficiencies and TODOs.
Don't use it in production.

The core addition of this patch is the HBA option "ldapmap", which
behaves like the existing "map" option except that the list of Postgres
roles is constructed from the results of an LDAP query. With this
method, role authorization can be controlled centrally, without needing
cron scripts for LDAP-to-Postgres synchronization.

The LDAP query must be in the form of an ldap[s]:// URL and it must
return exactly one entry (for the user that is logging in). That entry's
attribute values are used as the list of roles that an authenticated
user is allowed to log in as. For example, an LDAP admin could give any
authorized users one or more "postgresRole" attributes, and the DBA
could write a role query in pg_ident that looks like

  myldapmap  /^(.*)$  ldaps://example.com/dc=example,dc=com?postgresRole?sub?(uid=\1)

The ldapmap option can only be used for those auth methods that already
support the map option (cert, gss, etc.). The ldaptls, ldapbinddn, and
ldapbindpasswd options for the ldap auth method are now also supported
for any HBA lines that use ldapmap.

Possible remaining work (and known problems):

- It'd be nice to map roles based on LDAP group membership.

- pg_ident's regex substitution into a query URL is fairly unsafe and
  needs to be rethought. For now I've restricted the characters that can
  be substituted.

- The new LDAP tests open up TCP listen_addresses on the test server; I
  don't know if that's actually safe to do.

- Some of the options for specifying the LDAP server are in pg_ident
  (the scheme, host, and port), some are in pg_hba (ldaptls and the bind
  settings), and some are in the environment (client certificates, etc.)
  That is not easy to admin.

- Querying for multiple attributes is allowed, but only the first one
  returned by the server will be honored; the rest are silently ignored.

- The LDAP connection code might need to be deduplicated with
  InitializeLDAPConnection(), as they have substantial similarity.

- There's not much logging to help you when you're writing a role query,
  or to do a postmortem if something is wrong.

- Various others, called out inline...