ldapmap: implement SASL binding
Add the "ldapsaslmechs" HBA option, which tells the ldapmap query to use
SASL binding (instead of simple username/password binding) and specifies
the SASL mechanism list that is accepted. This allows the DBA to further
lock down the LDAP connection and even get rid of bind passwords
altogether.

This new feature is gated on the existence of the <sasl/sasl.h> header,
which we need in order to implement the SASL callback API. (No new
runtime dependencies are needed.)

The currently tested mechanisms are

- EXTERNAL, which authenticates using a TLS client certificate, and

- SCRAM-SHA-1, which performs mutual password authentication without
  ever sending the password over the connection. (Other SCRAM-*
  mechanisms should work too but I haven't tested them.)

TODOs:

- This seems like it would be useful for the ldap auth method, too.

- I reuse ldapbinddn for the SASL auth name, but it's probably not
  actually a DN in practice (and it's certainly not a DN for the test
  case I provided). Maybe make a new HBA option?